Move LinkCollector into nanoc-checking
