+++ nanoc-checking/Rakefile
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'rake/testtask'
 require 'rspec/core/rake_task'
 require 'rubocop/rake_task'
 
@@ -9,7 +10,13 @@ RSpec::Core::RakeTask.new(:spec) do |t|
   t.verbose = false
 end
 
-task test: :spec
+Rake::TestTask.new(:test_all) do |t|
+  t.test_files = Dir['test/**/test_*.rb']
+  t.libs << 'test'
+  t.verbose = false
+end
+
+task test: %i[spec test_all]
 
 task :gem do
   sh('gem build *.gemspec')
