+++ nanoc-checking/lib/nanoc/checking.rb
@@ -13,6 +13,7 @@ require_relative 'checking/check'
 require_relative 'checking/checks'
 require_relative 'checking/command_runners'
 require_relative 'checking/dsl'
+require_relative 'checking/link_collector'
 require_relative 'checking/runner'
 require_relative 'checking/loader'
 require_relative 'checking/issue'
