+++ nanoc-checking/lib/nanoc/checking/checks/external_links.rb
@@ -13,7 +13,7 @@ module Nanoc
           # Find all broken external hrefs
           # TODO: de-duplicate this (duplicated in internal links check)
           filenames = output_html_filenames.reject { |f| excluded_file?(f) }
-          hrefs_with_filenames = ::Nanoc::Extra::LinkCollector.new(filenames, :external).filenames_per_href
+          hrefs_with_filenames = ::Nanoc::Checking::LinkCollector.new(filenames, :external).filenames_per_href
           results = select_invalid(hrefs_with_filenames.keys.shuffle)
 
           # Report them
