+++ nanoc/nanoc.manifest
@@ -16,7 +16,6 @@ lib/nanoc/extra.rb
 lib/nanoc/extra/core_ext.rb
 lib/nanoc/extra/core_ext/time.rb
 lib/nanoc/extra/jruby_nokogiri_warner.rb
-lib/nanoc/extra/link_collector.rb
 lib/nanoc/extra/srcset_parser.rb
 lib/nanoc/filters.rb
 lib/nanoc/filters/asciidoc.rb
